Genericize Handler for event types.
Handler is now Handler[E], where E is Event or a subclass, and check() and
run() both take an E. The main loop in Bot will check whether each event is
compatible with a Handler's type before passing it to check() -- the Handler
never even hears about events of an incompatible type.

This saves us having to do a bunch of one-off isinstance checks on a
Handler-by-Handler basis. It also clears several type-checking errors now that
we're explicit about what we're doing.