Give every Event a "reply connection."
This replaces the reply callable, since it was always the source connection's
say() method anyway. Now, the response returned from a handler function is
passed to the reply connection's say(), but the reply connection need not be
the same as the source: for example, a TwitchWebookConnection generates
TwitchWebhookEvents with a TwitchChatConnection for replies -- with the effect
that chat messages can be triggered by webhook events.

Like the callable was, the connection is still optional. When the reply
connection is None, handlers must return None, since there's nowhere for a
message to go.