Split up Connection, adding ChatConnection. Add it to Event.
Now Connection only has `run` and `shutdown`, and ChatConnection extends it to
add `say`. This obviates the need for the NotImplementedError hackery also.

Rather than an optional reply callable, Events now carry an optional reply
ChatConnection. It doesn't need to be the same as the creating Connection (but
the creating Connection still needs to have a reference to it, so e.g.
TwitchWebhookConnection is now constructed with a TwitchChatConnection). This
improves on the old way, where the _Handler_ would have to know where to send
the reply, if it wasn't back to the source. It also paves the way for
ChatConnection to grow methods beyond `say`, e.g. for moderation.