Add subkeys to the database.
Previously everything in the DB was a key-value pair. Now, a key can either be
associated with a value, or with a set of subkey-value pairs. All keys,
subkeys, and values are strings.

This keeps the API simple enough to explain reasonably compactly, but offers a
little more support for structured data, so handlers don't have to do a lot of
string packing and unpacking.

The API still might change.