data: Make Namespace objects sharable between threads.
Previously Namespaces couldn't be shared because sqlite3.Connections
can't be. Now, instead of each Namespace wrapping a Connection, it wraps
a collection of thread-local Connections.

The Namespace's external behavior is unchanged, since each method acted
atomically on the DB already. If that changes in the future, this
behavior could be confusing -- but in principle that was already the
case, since there was never a guarantee that each Namespace has
exclusive access to its data; two identical Namespace objects can exist.